Minifying
Merged redundancies
